refactor(maps): extract router helper cluster to _router_helpers.py
Move the 12 response-builder / access-check helpers (~265 lines) out of maps/router.py into a sibling _router_helpers.py module (they have no router dependency, so no circular import). Shrinks router.py 2107 -> 1844 lines and separates pure helpers from endpoint definitions. Lowers the Phase 276 CODE-01 LOC allowlist from the 2150 band-aid to 1900. Routes unchanged (389 total); further reduction below the 1500 default is queued as a share/media/layers sub-router split (Phase 226/238 pattern).
